+++ armada_backend/api_create.py
@@ -1,13 +1,11 @@
 import base64
 import json
-import traceback
 
 import web
 
-import api_base
-import docker_client
-from armada_backend.utils import get_logger
+from armada_backend import api_base, docker_client
 from armada_backend.api_run_hermes import process_hermes
+from armada_backend.utils import get_logger
 from armada_command.armada_utils import split_image_path
 from armada_command.dockyard.alias import INSECURE_REGISTRY_ERROR_MSG
 
@@ -23,7 +21,7 @@ class Create(api_base.ApiCommand):
             raise ValueError('Field run_command cannot be empty.')
 
         if kwargs:
-            get_logger().warning('JSON data sent to API contains unrecognized keys: {}'.format(list(kwargs.keys())))
+            get_logger().warning('JSON data sent to API contains unrecognized keys: %s', list(kwargs.keys()))
 
         # Set default values:
         environment = environment or {}
@@ -118,10 +116,10 @@ class Create(api_base.ApiCommand):
                     logged_in = True
                     break
                 except Exception as e:
+                    get_logger().debug(e)
                     login_exceptions.append(e)
+
             if not logged_in:
-                for e in login_exceptions:
-                    traceback.print_tb(e.__traceback__)
                 raise login_exceptions[0]
 
     def _get_docker_api(self, dockyard_address, dockyard_user, dockyard_password):
@@ -156,9 +154,8 @@ class Create(api_base.ApiCommand):
     def POST(self):
         try:
             post_data = json.loads(web.data())
-        except:
-            traceback.print_exc()
-            return self.status_error('API Run: Invalid input JSON.')
+        except Exception as e:
+            return self.status_exception('API Run: Invalid input JSON.', e)
 
         try:
             long_container_id = self._create_service(**post_data)