+++ armada_backend/api_ship.py
@@ -2,17 +2,17 @@ import json
 import logging
 import os
 import time
-import xmlrpclib
 from socket import gethostname
 
-import api_base
-import consul_config
-from armada_backend.utils import deregister_services, set_ship_name, get_logger
+import xmlrpclib
+
+from armada_backend import api_base, consul_config
+from armada_backend.runtime_settings import override_runtime_settings
+from armada_backend.utils import deregister_services, get_current_datacenter, get_logger, get_other_ship_ips, \
+    set_ship_name, get_ship_name
 from armada_command import armada_api
-from armada_command.consul.consul import consul_query, consul_put
-from runtime_settings import override_runtime_settings
-from utils import get_ship_name, get_other_ship_ips, get_current_datacenter
 from armada_command.consul import kv
+from armada_command.consul.consul import consul_query, consul_put
 
 
 def _get_current_consul_mode():
@@ -114,7 +114,7 @@ class Join(api_base.ApiCommand):
         if _restart_consul():
             supervisor_server = xmlrpclib.Server('http://localhost:9001/RPC2')
             hermes_init_output = supervisor_server.supervisor.startProcessGroup('hermes_init')
-            get_logger().info('hermes_init start: {}'.format(hermes_init_output))
+            get_logger().info('hermes_init start: %s', hermes_init_output)
             set_ship_name(ship)
             for key, data in local_services_data.items():
                 kv.kv_set(key, data)