+++ armada_backend/api_start.py
@@ -1,10 +1,7 @@
-import traceback
-
-import api_base
-import docker_client
+from armada_backend import api_base, docker_client
+from armada_backend.utils import shorten_container_id, get_ship_name
 from armada_command.consul.consul import consul_query
 from armada_command.consul.kv import save_service
-from utils import shorten_container_id, get_ship_name
 
 
 class Start(api_base.ApiCommand):
@@ -35,5 +32,4 @@ class Start(api_base.ApiCommand):
             service_endpoints = self._start_container(long_container_id)
             return self.status_ok({'endpoints': service_endpoints})
         except Exception as e:
-            traceback.print_exc()
             return self.status_exception("Cannot start service's container", e)